+++ Makefile
@@ -31,7 +31,7 @@ test::
 
 clean::
 	rm -rf $(TOPDIR)/bin
-	rm -rf $(TOPDIR)/internal/frontend/build
+	rm -rf $(TOPDIR)/frontend_server/build
 	rm -rf $(TOPDIR)/frontend/out
 	rm -rf $(TOPDIR)/frontend/.next
 
@@ -103,14 +103,14 @@ $(TOPDIR)/bin: $(TOPDIR)
 	mkdir -p $(TOPDIR)/bin
 
 # Linux AMD64 Builds
-$(TOPDIR)/bin/amd64-linux/habitat: $(TOPDIR)/bin internal/frontend/build
+$(TOPDIR)/bin/amd64-linux/habitat: $(TOPDIR)/bin frontend_server/build
 	GOARCH=amd64 GOOS=linux go build -o $(TOPDIR)/bin/amd64-linux/habitat $(TOPDIR)/cmd/node/main.go
 
 $(TOPDIR)/bin/amd64-linux/habitat-amd64-linux.tar.gz: $(TOPDIR)/bin/amd64-linux/habitat
 	tar -czf $(TOPDIR)/bin/amd64-linux/habitat-amd64-linux.tar.gz -C $(TOPDIR)/bin/amd64-linux habitat
 
 # Darwin AMD64 Builds
-$(TOPDIR)/bin/amd64-darwin/habitat: $(TOPDIR)/bin internal/frontend/build
+$(TOPDIR)/bin/amd64-darwin/habitat: $(TOPDIR)/bin frontend_server/build
 	GOARCH=amd64 GOOS=darwin go build -o $(TOPDIR)/bin/amd64-darwin/habitat $(TOPDIR)/cmd/node/main.go
 
 $(TOPDIR)/bin/amd64-darwin/habitat-amd64-darwin.tar.gz: $(TOPDIR)/bin/amd64-darwin/habitat
@@ -131,7 +131,7 @@ frontend-types: frontend/types/api.ts
 PHONY += frontend-types
 
 # Embed the frontend in the binary
-internal/frontend/build: frontend/types/api.ts
+frontend_server/build: frontend/types/api.ts
 	cd $(TOPDIR)/frontend && pnpm install && pnpm run build
-	mkdir -p $(TOPDIR)/internal/frontend/build
-	cp -r $(TOPDIR)/frontend/out/* $(TOPDIR)/internal/frontend/build
+	mkdir -p $(TOPDIR)/frontend_server/build
+	cp -r $(TOPDIR)/frontend/out/* $(TOPDIR)/frontend_server/build
